move agent label map factories out of icon.tsx
